credential_provider/gcp/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opentelemetry-instrumentation-credential-provider-component"
dynamic = ["version"]
description = "OpenTelemetry OTLP exporter credential provider"
readme = "README.rst"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
  { name = "OpenTelemetry Authors", email = "cncf-opentelemetry-contributors@lists.cncf.io" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "google-auth ~= 2.38",
  "grpcio ~= 1.7",
]

[project.entry-points.opentelemetry_otlp_credential_provider]
google_gprc_credential_provder = "opentelemetry.credentials.gcp:channel_credentials_provider"
google_http_credential_provder = "opentelemetry.credentials.gcp:session_provider"

[tool.hatch.version]
path = "src/opentelemetry/credentials/gcp/version.py"

[tool.hatch.build.targets.sdist]
include = [
  "/src",
]

[tool.hatch.build.targets.wheel]
packages = ["src/opentelemetry"]